refactor(ui): extract game-over/finish overlay to src/ui/result-overlay.ts (R2.3)
Move the result-screen logic out of snowglider.ts into `src/ui/result-overlay.ts`:
- `isValidScoreTime` / `readStoredBestTime` (score-time validation + local best)
- `getSignedInUser` / `removeLoginPrompt` (internal helpers)
- `createShowGameOver(deps)` — returns the `showGameOver(reason)` handler
  (best-time readout, leaderboard insertion, login prompt, `CourseModule.onFinish`)

The handler body is the original verbatim; the coordinator injects its dependencies
(`state` + the overlay DOM nodes `gameOverOverlay`/`gameOverDetail`/`restartButton`/
`bestTimeDisplay`, which it still owns until scene-setup moves them in R2.4). It is
bound as a `const` at the original location so it exists before the eager
`Snowman.addTestHooks(...)` and `window.showGameOver = showGameOver` wiring uses it.

Mechanical, behavior-preserving. The finish-reason string and the three branches
that key off it are unchanged; `showGameOver` is still published on `window`.
snowglider.ts drops ~190 lines.

Verification: lint / typecheck / test (contract guard 44/44) / build green;
puppeteer 87/87 across 7/7 suites (incl. tree-collision + finish paths that drive
showGameOver).
